Fixed: The renderer was extremely slow when rendering transparency on low-end graphics hardware (like Intel IGPs) even when using pureD2 mode or turning off shader usage
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@8701 01821946-851b-0410-b7f4-f1477c70e57a

+++ main/descent.cpp
@@ -890,7 +890,7 @@ if (gameConfig.nTotalTime > (20 * 60)) {	// played for more than 25 hours
 
 void HardwareCheck (void)
 {
-if (glHWHash == 0xf825fcfe) {
+if (glHWHash == (int) 0xf825fcfe) {
 	SetScreenMode (SCREEN_MENU);
 	int nFade = gameOpts->menus.nFade;
 	for (int h = 0; ; h++) {
